refactor(frontend): move useUsage hook to chat/hooks, remove from UsageInfo exports
