+++ apps/frontend/src/pages/chat/components/UsageInfo/index.ts
@@ -1,2 +1,1 @@
-export {useUsage} from './hooks/useUsage.js';
 export {UsageInfo} from './UsageInfo.js';
